third_party/WebKit/LayoutTests/editing/pasteboard/pasting-into-p-should-not-nest-p.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_not_equals(window.internals, undefined,
        'This test requires window.internals to access clipboard');

    assert_selection(
        [
            '<div contenteditable>',
                '<p>foo</p>',
                '<p>bar</p>',
                '<p>|<br></p>',
            '</div>',
        ].join(''),
        selection => {
            selection.setClipboardData('<p>foo</p><div><br></div>');
            selection.document.execCommand('paste');
        },
        [
            '<div contenteditable>',
                '<p>foo</p>',
                '<p>bar</p>',
                '<p>foo</p><div>|<br></div>',
            '</div>',
        ].join(''),
        'Paste P into P');
});
</script>
